PARTIAL COMMIT: new async_io infrastructure
The idea is that I want to cut libevent out of WashingtonDC because it's the
root-cause of way too many build problems, and I don't really need it because
it's made for programs like webservers that have much more demanding scaling
needs (in terms of socket connections) than WashingtonDC does.  For my purposes
select ought to be good enough.