sh4: get rid of Sh4OpArgs
One problem is it might not be portable to non-x86 hosts because it relies upon
C bitfields perfectly overlaying on top of SH-4 instructions.

Another problem is that it forces me to have sh4-specific code in the jit, and
that won't do at all when I add in the ARM7 jit code.

The new implementation represents instructions as integers, so the only concern
is what type of int to use.  Either uint32_t or unsigned int will be
appropriate, although IDK which one is better.