jit: x86_64: don't call evict_register in emit_jump
move_slot_to_reg will take care of the existing contents of the register on its
own.  Calling evict_register was counter productive because it was causing
useless register shuffling in situations where the slot was already in the
correct register because evict_register would move the slot to a different
register and then move_slot_to_reg would move it back to the original register.